LLM_Langchain_BOT/requirements.txt
langchain==0.3.9
langchainhub==0.1.20
langchain-community==0.3.8
langchain_core==0.3.21
langchain_openai==0.2.8
langchain_text_splitters==0.3.2
openai==1.54.4
SQLAlchemy==2.0.35
streamlit==1.40.1
duckduckgo-search==7.2.0
pypdf==5.1.0
fastembed==0.4.2
docarray==0.40.0
validators==0.34.0
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==11.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3
